Phase 3: Fix type consistency in backward pass - use double* throughout
TYPE FIXES:
- Fixed gradient buffer types from float* to double* throughout backward pass
- Updated cllm_backward_training signature to use double* gradient_buffer
- Fixed attention_backward_full to use double* parameters
- Fixed softmax_backward to use double* parameters
- Updated all local gradient variables to use double*
- Updated header declaration in cllm_training.h

CONSISTENCY ACHIEVED:
- Model weights: double* (64-bit)
- Training gradients: double* (64-bit)
- Backward pass buffers: double* (64-bit)
- Layer operations: float* (32-bit for performance)

BACKWARD PASS STATUS:
✅ Already fully implemented with:
  - Gradient computation for loss w.r.t. logits
  - Backward through output projection
  - Loop through ALL layers (top to bottom)
  - Layer norm backward pass
  - Attention backward pass (full and simplified)
  - Feedforward backward pass
  - All gradients computed (100%, not just 7.5%)

BUILD STATUS: ✅ Success with warnings only
NEXT: Test training to verify no NaN gradients

+++ include/cllm_training.h
@@ -198,6 +198,6 @@ void cllm_training_cleanup(CLLMTraining* training);
 
 /* Training step functions (used by multi-threading and production features) */
 float cllm_forward_training(CLLMTraining* training, uint32_t* input_tokens);
-void cllm_backward_training(CLLMTraining* training, uint32_t* target_tokens, float* gradient_buffer);
+void cllm_backward_training(CLLMTraining* training, uint32_t* target_tokens, double* gradient_buffer);
 
 #endif /* CLLM_TRAINING_H */
